tests(security form) repair unit test
