+++ src/components/Dashboard/SettingsSection/SettingsContent/SecurityForm/components/PasswordFields/__tests__/PasswordFields.test.tsx
@@ -1,51 +1,38 @@
- import { describe, it, expect, vi } from 'vitest';
-import { renderHook } from '@testing-library/react';
-import { useSecurityToasts } from '../useSecurityToasts';
-import { SecurityError } from '../../../utils/api/security';
-import { useToast } from '../../../contexts/ToastContext';
+import React from 'react';
+import { describe, it, expect, vi } from 'vitest';
+import { render, screen } from '@testing-library/react';
+import { PasswordFields } from '../PasswordFields.component';
 
-vi.mock('../../../contexts/ToastContext');
+describe('PasswordFields', () => {
+  const mockForm = {
+    register: vi.fn(),
+  };
 
-describe('useSecurityToasts', () => {
-  const mockShowToast = vi.fn();
+  const mockErrors = {
+    currentPassword: undefined,
+    newPassword: undefined,
+    confirmPassword: undefined
+  };
 
-  beforeEach(() => {
-    vi.clearAllMocks();
-    (useToast as jest.Mock).mockReturnValue({ showToast: mockShowToast });
+  it('renders all password fields', () => {
+    render(<PasswordFields form={mockForm as any} errors={mockErrors} />);
+    
+    expect(screen.getByPlaceholderText('Wprowadź aktualne hasło')).toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Wprowadź nowe hasło')).toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Powtórz nowe hasło')).toBeInTheDocument();
   });
 
-  it('handles success correctly', () => {
-    const { result } = renderHook(() => useSecurityToasts());
-    
-    result.current.handleSuccess();
-    expect(mockShowToast).toHaveBeenCalledWith('Hasło zostało zmienione', 'success');
-  });
+  it('displays error messages when present', () => {
+    const errorsWithMessages = {
+      currentPassword: { message: 'Wymagane pole' },
+      newPassword: { message: 'Hasło jest za krótkie' },
+      confirmPassword: { message: 'Hasła nie są identyczne' }
+    };
 
-  it('handles invalid password error correctly', () => {
-    const { result } = renderHook(() => useSecurityToasts());
+    render(<PasswordFields form={mockForm as any} errors={errorsWithMessages} />);
     
-    result.current.handleError(new SecurityError('INVALID_CURRENT_PASSWORD'));
-    expect(mockShowToast).toHaveBeenCalledWith('Aktualne hasło jest nieprawidłowe', 'error');
-  });
-
-  it('handles general error correctly', () => {
-    const { result } = renderHook(() => useSecurityToasts());
-    
-    result.current.handleError(new Error('Unknown error'));
-    expect(mockShowToast).toHaveBeenCalledWith('Wystąpił błąd podczas zmiany hasła', 'error');
-  });
-
-  it('handles cancel correctly', () => {
-    const { result } = renderHook(() => useSecurityToasts());
-    
-    result.current.handleCancel();
-    expect(mockShowToast).toHaveBeenCalledWith('Zmiany zostały anulowane', 'success');
-  });
-
-  it('handles cancel error correctly', () => {
-    const { result } = renderHook(() => useSecurityToasts());
-    
-    result.current.handleCancelError();
-    expect(mockShowToast).toHaveBeenCalledWith('Nie udało się anulować zmian', 'error');
+    expect(screen.getByText('Wymagane pole')).toBeInTheDocument();
+    expect(screen.getByText('Hasło jest za krótkie')).toBeInTheDocument();
+    expect(screen.getByText('Hasła nie są identyczne')).toBeInTheDocument();
   });
 });
